updated things to new osap
