+++ arduino/oled-thing/oled-thing.ino
@@ -1,8 +1,4 @@
 #include <osap.h>
-#include <vt_endpoint.h>
-#include <vp_arduinoSerial.h>
-#include <core/ts.h>
-#include <Wire.h>
 #include <Adafruit_GFX.h>
 #include <Adafruit_SSD1306.h>
 
@@ -16,17 +12,11 @@
 #define SCREEN_ADDRESS 0x3C
 Adafruit_SSD1306 display(SCREEN_WIDTH, SCREEN_HEIGHT);
 
-// message-passing memory allocation 
-#define OSAP_STACK_SIZE 10
-VPacket messageStack[OSAP_STACK_SIZE];
-// type of board (firmware name)
-OSAP osap("oled", messageStack, OSAP_STACK_SIZE);
+OSAP_Runtime osap;
+OSAP_Gateway_USBSerial serLink(&Serial);
+OSAP_Port_DeviceNames namePort("oled");
 
-// ---------------------------------------------- 0th Vertex: OSAP USB Serial
-VPort_ArduinoSerial vp_arduinoSerial(&osap, "usbSerial", &Serial);
-
-// ---------------------------------------------- 1th Vertex: String input Endpoint 
-EP_ONDATA_RESPONSES onStringData(uint8_t* data, uint16_t len) {
+void setString(uint8_t* data, uint16_t len) {
   // first byte is the text size
   uint8_t txt_size = data[0];
   // the rest is the text
@@ -40,15 +30,12 @@ EP_ONDATA_RESPONSES onStringData(uint8_t* data, uint16_t len) {
   display.setTextSize(txt_size);
   display.print(txt);
   display.display();
-
-  return EP_ONDATA_ACCEPT;
 }
 
-Endpoint stringEndpoint(&osap, "stringEndpoint", onStringData);
+OSAP_Port_Named setString_port("setString", setString);
 
 void setup() {
-  osap.init();
-  vp_arduinoSerial.begin();
+  osap.begin();
 
   display.begin(SSD1306_SWITCHCAPVCC, SCREEN_ADDRESS);
