+++ arduino/potentiometer-thing/potentiometer-thing.ino
@@ -1,39 +1,28 @@
 #include <osap.h>
-#include <vt_endpoint.h>
-#include <vp_arduinoSerial.h>
 
 #define PIN_POT1 8
 #define PIN_POT2 7
 
-// message-passing memory allocation 
-#define OSAP_STACK_SIZE 10
-VPacket messageStack[OSAP_STACK_SIZE];
-// type of board (firmware name)
-OSAP osap("potentiometer", messageStack, OSAP_STACK_SIZE);
+OSAP_Runtime osap;
+OSAP_Gateway_USBSerial serLink(&Serial);
+OSAP_Port_DeviceNames namePort("potentiometer");
 
-// ---------------------------------------------- 0th Vertex: OSAP USB Serial
-VPort_ArduinoSerial vp_arduinoSerial(&osap, "usbSerial", &Serial);
-
-// ---------------------------------------------- 1 Vertex
-boolean prePotQuery(void);
-
-Endpoint tofEndpoint(&osap, "potentiometerQuery", prePotQuery);
-
-boolean prePotQuery(void) {
-  uint8_t buf[4];
+size_t readPotentiometer(uint8_t* data, size_t len, uint8_t* reply) {
   uint16_t value1 = analogRead(PIN_POT1);
   uint16_t value2 = analogRead(PIN_POT2);
-  buf[0] = value1 & 0xFF;
-  buf[1] = value1 >> 8 & 0xFF;
-  buf[2] = value2 & 0xFF;
-  buf[3] = value2 >> 8 & 0xFF;
-  tofEndpoint.write(buf, 4);
-  return true;
+
+  reply[0] = value1 & 0xFF;
+  reply[1] = value1 >> 8 & 0xFF;
+  reply[2] = value2 & 0xFF;
+  reply[3] = value2 >> 8 & 0xFF;
+
+  return 4;
 }
 
+OSAP_Port_Named readPotentiometer_port("readPotentiometer", readPotentiometer);
+
 void setup() {
-  osap.init();
-  vp_arduinoSerial.begin();
+  osap.begin();
 
   pinMode(PIN_POT1, INPUT);
   pinMode(PIN_POT2, INPUT);
