--- src/lib/virtualThings/capacitive_new.js
@@ -1,54 +1,0 @@
-import { osap } from "../osapjs/osap";
-import { readUint16 } from "../osapjs/utils/serializers.js"
-
-let N_PADS = 6
-
-export default function(name) {
-
-  return {
-    setRGB: async (r, g, b) => {
-      try {
-        const datagram = new Uint8Array([
-          255 - r * 255,
-          255 - g * 255,
-          255 - (b * 255) / 2
-        ])
-
-        await osap.send(name, "setRGB", datagram);
-      } catch (err) {
-        console.error(err)
-      }
-    },
-    readPad: async (index) => {
-      try {
-        // just get 'em all
-        const data = await osap.send(name, "readPad", new Uint8Array([]));
-        const vals = []
-        for(let p = 0; p < N_PADS; p ++){
-          vals.push(readUint16(data, p * 2))
-        }
-
-        return vals[index]/1023
-      } catch (err) {
-        console.error(err)
-      }
-    },
-    api: [
-      {
-        name: "setRGB",
-        args: [
-          "red: 0 to 1",
-          "green: 0 to 1",
-          "blue: 0 to 1"
-        ]
-      },
-      {
-        name: "readPad",
-        args: [
-          "index: int 0 to 5"
-        ],
-        return: "0 to 1"
-      },
-    ]
-  }
-}
