--- src/lib/virtualThings/mosfet_new.js
@@ -1,24 +1,0 @@
-import { osap } from "../osapjs/osap";
-
-export default function mosfetThing(name) {
-
-  return {
-    setGate: async (value) => {
-      try {
-        const datagram = new Uint8Array([ 255 * value ])
-        await osap.send(name, "setGate", datagram);
-
-      } catch (err) {
-        console.error(err)
-      }
-    },
-    api: [
-      {
-        name: "setGate",
-        args: [
-          "value: 0 to 1"
-        ],
-      }
-    ]
-  }
-}
