+++ src/lib/virtualThings/servo.js
@@ -1,36 +1,8 @@
-/*
-rgbbThing.js
+import { osap } from "../osapjs/osap";
+import Serializers from "../osapjs/utils/serializers"
+const writeUint16 = Serializers.writeUint16;
 
-a "virtual thing" - of course 
-
-Jake Read, Leo McElroy and Quentin Bolsee at the Center for Bits and Atoms
-(c) Massachusetts Institute of Technology 2022
-
-This work may be reproduced, modified, distributed, performed, and
-displayed for any purpose, but must acknowledge the open systems assembly protocol (OSAP) and modular-things projects.
-Copyright is retained and must be preserved. The work is provided as is;
-no warranty is provided, and users accept all liability.
-*/
-
-import { TS } from "../osapjs/core/ts.js"
-import PK from "../osapjs/core/packets.js"
-
-export default function servo(osap, vt, name) {
-  // ---------------------------------- OSAP... stuff, 
-  let routeToFirmware = PK.VC2VMRoute(vt.route)
-
-  // this is the '1th' vertex, so we address it like-this:
-  let servoPulseWidthMirror = osap.endpoint("servoPulseWidthMirror")
-  servoPulseWidthMirror.addRoute(PK.route(routeToFirmware).sib(1).end())
-
-  // we should have a setup function:
-  const setup = async () => {
-    try {
-      // noop 
-    } catch (err) {
-      throw err
-    }
-  }
+export default function servo(name) {
 
   // calibrated-angle-bounds, 
   let pulseBounds = [1000, 2000] // pulse-width bounds
@@ -38,10 +10,10 @@ export default function servo(osap, vt, name) {
 
   let writeMicroseconds = async (us) => {
     try {
-      us = Math.round(us)
-      let datagram = new Uint8Array(2)
-      TS.write('uint16', us, datagram, 0)
-      await servoPulseWidthMirror.write(datagram, "acked")
+      us = Math.round(us);
+      let datagram = new Uint8Array(2);
+      writeUint16(datagram, 0, us);
+      await osap.send(name, "writeMicroseconds", datagram);
     } catch (err) {
       throw err
     }
@@ -73,8 +45,6 @@ export default function servo(osap, vt, name) {
     writeMicroseconds,
     writeAngle,
     setCalibration,
-    setup,
-    vt,
     api: [
       {
         name: "writeMicroseconds",
