+++ src/lib/virtualThings/timeOfFlight.js
@@ -1,40 +1,19 @@
-/*
-a "virtual thing" - of course 
+import { osap } from "../osapjs/osap";
+import Serializers from "../osapjs/utils/serializers"
+const readUint16 = Serializers.readUint16;
 
-Jake Read, Leo McElroy and Quentin Bolsee at the Center for Bits and Atoms
-(c) Massachusetts Institute of Technology 2022
-
-This work may be reproduced, modified, distributed, performed, and
-displayed for any purpose, but must acknowledge the open systems assembly protocol (OSAP) and modular-things projects.
-Copyright is retained and must be preserved. The work is provided as is;
-no warranty is provided, and users accept all liability.
-*/
-
-import { TS } from "../osapjs/core/ts.js"
-import PK from "../osapjs/core/packets.js"
-
-export default function(osap, vt, name) {
-
-  let routeToFirmware = PK.VC2VMRoute(vt.route);
-  let tofQuery = osap.query(PK.route(routeToFirmware).sib(1).end());
-  // why not just an endpoint?
-
-  const setup = async () => { }
+export default function(name) {
 
   return {
     readDistance: async () => {
       try {
-        const data = await tofQuery.pull();
-        // could be
-        // data.read("uint16");
-        const val = TS.read("uint16", data, 0);
+        const data = osap.send(name, "readDistance");
+        const val = readUint16(data, 0);
         return val;
       } catch (err) {
         console.error(err)
       }
     },
-    setup,
-    vt,
     api: [
       {
         name: "readDistance",
